Pod and runtime configuration.
Including fixes for recover from conflict, various fixes for issues with updates of the different configs.
Not craeting deployment immediately in RunnerActions, create or update is triggered through app input secret.
Adding a timestamp label to the input secret so it changes with every deployment.
Updated operator image for this commit.

Fixing up tests.