Fix an issue detecting a CRD version.
The previous iteration of this code worked on the assumption that there
is a singluar `version` field in the CRD, this is not true in newer
versions of K8s. This version extracts all the `versions[].name`'s in
the cluster and checks if any of those versions matches the desiered
one. The desired version have to be the stored version, if we write a
supported, but not current version of the CR, it will get converted and
usable by the operator.

Have also added a better error message when the plugin cannot detect a
version, this is most likely to happen when you have a older cluster.