fix: bumping up version (#1108)
As per documentation https://github.com/strimzi/strimzi-kafka-operator/releases with the new release the Kafka CR doesn't work no more with version 2.7.1. Deploy a CR with that version raises and exception that states that 2.7.1 is not supported anymore and you need to chose among 2.8.0, 2.8.1 and 3.0.0.
I have choosen 3.0.0 only to be the latest, no other reason.